Feature and BugFix
Melhorando desempenho de query's e consertando criação de vendas

+++ src/main/java/com/aa/algoritmo/domain/model/SalesItem.java
@@ -1,10 +1,11 @@
 package com.aa.algoritmo.domain.model;
 
+import java.math.BigDecimal;
+
 public record SalesItem(
         Integer id,
         Integer productId,
         String productName,
         Integer quantity,
-        Integer price
-) {
-}
+        BigDecimal price
+) {}
